2019-9-16   20:11    --》kong

+++ src/main/java/com/xdmd/IntranetEnvironment/contractmanage/pojo/KeyResearchDevelopersDTO.java
@@ -41,7 +41,7 @@ public class KeyResearchDevelopersDTO {
     private String workTask;
 
     @ApiModelProperty("为本课题工作时间（%）")
-    private Integer workingTime;
+    private Double workingTime;
 
     public KeyResearchDevelopersDTO() {
     }
